requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.30.0
opencv-python-headless>=4.8.0
numpy>=1.23.0
Pillow>=9.5.0
